requirements.txt
Pillow
opencv-python
numpy
python3_xlib
pyautogui
pywin32
ipython
pytesseract
coverage
keyboard==0.11.0
GitPython
Django
channels
asgiref
channels_redis
pytz
requests